requirements/base.txt
dj-database-url==0.4.0
Django==1.8.13
gunicorn==19.4.5
whitenoise==2.0.6
django-extended-choices==1.0.7
django-extensions==1.6.1
ipython==4.1.2
Werkzeug==0.11.4
djangorestframework==3.3.3
Markdown==2.6.6
django-cors-headers==1.1.0

# Django rest swagger
PyYAML==3.11
argh==0.23.2
argparse==1.2.1
jsonschema==2.5
docutils>=0.11
